refactor: redisign components
