+++ src/componentsNew/RadioControl/RadioControl.module.scss
@@ -1,0 +1,24 @@
+.radioWrapper {
+  display: flex;
+  flex-direction: column;
+  gap: 11px;
+  margin-bottom: 20px;
+
+  @media screen and (max-width: 767px) {
+    margin-bottom: 15px;
+    flex-direction: row;
+  }
+}
+
+.radioLabel {
+  font: var(--font-text-6);
+  color: var(--primary-border);
+  text-transform: uppercase;
+  margin-bottom: 11px;
+}
+
+.sorting {
+  display: flex;
+  flex-direction: column;
+  gap: 8px;
+}
