+++ src/componentsNew/SortControl/SortControl.module.scss
@@ -1,0 +1,18 @@
+.sortingWrapper {
+  display: flex;
+  flex-flow: row wrap;
+  gap: 36.5px;
+  row-gap: 12px;
+  margin-top: 10px;
+
+  @media screen and (max-width: 767px) {
+    justify-content: center;
+    max-width: 340px;
+  }
+}
+
+.sorting {
+  display: flex;
+  flex-direction: column;
+  gap: 8px;
+}
