+++ src/pages/BorrowPage/components/BorrowManual/BorrowManual.module.scss
@@ -1,26 +1,7 @@
 @import '../../../../scss/abstracts/mixins';
 
-.header {
-  display: flex;
-  justify-content: center;
-  margin-bottom: 48px;
-
-  @media screen and (max-width: 1023px) {
-    margin-top: 30px;
-    flex-direction: column;
-  }
-}
-
 .headerActive {
   padding-right: 302px;
-
-  @media (max-width: 767px) {
-    padding-right: 0;
-  }
-}
-
-.filtersActive {
-  padding-right: 334px;
 
   @media (max-width: 767px) {
     padding-right: 0;
@@ -38,26 +19,6 @@
   text-align: center;
 }
 
-.sortWrapper {
-  display: flex;
-  justify-content: space-between;
-  flex-wrap: wrap;
-  margin: 20px 0;
-  position: relative;
-  width: 100%;
-  flex-shrink: 0;
-}
-
-.searchInput {
-  max-width: 173px;
-  margin-right: 20px;
-  width: unset;
-
-  @media screen and (max-width: 468px) {
-    margin-bottom: 20px;
-  }
-}
-
 .loader {
   display: flex;
   justify-content: center;
@@ -68,59 +29,10 @@
   min-width: 160px;
 }
 
-.sortingWrapper {
-  display: flex;
-  flex-flow: row wrap;
-  gap: 36.5px;
-  row-gap: 12px;
-  margin-top: 10px;
-
-  @media screen and (max-width: 767px) {
-    justify-content: center;
-    max-width: 340px;
-  }
-}
-
-.sorting {
-  display: flex;
-  flex-direction: column;
-  gap: 8px;
-}
-
 .label {
   color: var(--gray-color-2);
   font: var(--font-text-6);
   text-transform: uppercase;
-}
-
-.filtersDropdown {
-  position: absolute;
-  top: 45px;
-  right: 0;
-  width: 308px;
-
-  @media screen and (max-width: 767px) {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    position: fixed;
-    bottom: 0;
-    left: 0;
-    right: 0;
-    top: unset;
-    width: 100% !important;
-  }
-}
-
-.filters {
-  position: absolute;
-  top: 0;
-  right: 0;
-}
-
-.search {
-  margin-bottom: 32px;
-  max-width: 350px;
 }
 
 .connectBtn {
@@ -177,30 +89,5 @@
     animation: dots 2s linear infinite;
     width: 16px;
     content: '...';
-  }
-}
-
-.btnBack {
-  @include flex(row, nowrap, center, center);
-
-  background-color: var(--primary-background);
-  border: 1px solid var(--primary-border);
-  border-radius: 50%;
-  cursor: pointer;
-
-  position: absolute;
-  top: 110px;
-  left: 32px;
-  height: 40px;
-  width: 40px;
-
-  svg {
-    path {
-      fill: var(--secondary-border);
-    }
-  }
-
-  &:hover {
-    border-color: var(--secondary-border);
   }
 }
