+++ src/pages/BorrowPage/components/Sort/Sort.module.scss
@@ -1,0 +1,52 @@
+.sortWrapper {
+  display: flex;
+  justify-content: space-between;
+  flex-wrap: wrap;
+  margin: 20px 0;
+  position: relative;
+  width: 100%;
+  flex-shrink: 0;
+}
+
+.searchInput {
+  max-width: 173px;
+  margin-right: 20px;
+  width: unset;
+
+  @media screen and (max-width: 468px) {
+    margin-bottom: 20px;
+  }
+}
+
+.filtersDropdown {
+  position: absolute;
+  top: 45px;
+  right: 0;
+  width: 308px;
+
+  @media screen and (max-width: 767px) {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    position: fixed;
+    bottom: 0;
+    left: 0;
+    right: 0;
+    top: unset;
+    width: 100% !important;
+  }
+}
+
+.filters {
+  position: absolute;
+  top: 0;
+  right: 0;
+}
+
+.filtersActive {
+  padding-right: 334px;
+
+  @media (max-width: 767px) {
+    padding-right: 0;
+  }
+}
