+++ src/pages/BorrowPage/components/Sort/index.ts
@@ -1,0 +1,1 @@
+export { default } from './Sort';
